+++ packages/browser/types/index.d.ts
@@ -1,5 +1,5 @@
-import { History, LocationDetails, PartialLocation, HickoryLocation, AnyLocation, Options as RootOptions } from "@hickory/root";
-export { History, HickoryLocation, PartialLocation, AnyLocation, LocationDetails };
+import { History, LocationComponents, PartialLocation, SessionLocation, AnyLocation, Location, Options as RootOptions } from "@hickory/root";
+export { History, SessionLocation, PartialLocation, AnyLocation, Location, LocationComponents };
 export interface Options<Q> extends RootOptions<Q> {
     raw?: (pathname: string) => string;
 }