+++ packages/root/types/types/hickory.d.ts
@@ -1,4 +1,4 @@
-import { HickoryLocation, AnyLocation, PartialLocation } from "./location";
+import { SessionLocation, AnyLocation, PartialLocation } from "./location";
 import { LocationFactoryOptions, LocationMethods } from "./locationFactory";
 import { KeyMethods } from "./keygen";
 import { ConfirmationFunction, ConfirmationMethods } from "./navigationConfirmation";
@@ -6,7 +6,7 @@ export declare type ToArgument<Q> = string | PartialLocation<Q>;
 export declare type Action = "push" | "replace" | "pop";
 export declare type NavType = "anchor" | "push" | "replace";
 export interface PendingNavigation<Q> {
-    location: HickoryLocation<Q>;
+    location: SessionLocation<Q>;
     action: Action;
     finish(): void;
     cancel(nextAction?: Action): void;
@@ -14,7 +14,7 @@ export interface PendingNavigation<Q> {
 }
 export declare type ResponseHandler<Q> = (resp: PendingNavigation<Q>) => void;
 export interface History<Q> {
-    location: HickoryLocation<Q>;
+    location: SessionLocation<Q>;
     action: Action;
     toHref(to: AnyLocation<Q>): string;
     respondWith(fn: ResponseHandler<Q>): void;