+++ packages/root/types/types/location.d.ts
@@ -1,15 +1,15 @@
-export interface LocationDetails<Q> {
+export interface LocationComponents<Q> {
     pathname: string;
     query: Q;
     hash: string;
     state?: any;
 }
-export declare type PartialLocation<Q> = Partial<LocationDetails<Q>>;
-export interface KeylessLocation<Q> extends LocationDetails<Q> {
+export declare type PartialLocation<Q> = Partial<LocationComponents<Q>>;
+export interface Location<Q> extends LocationComponents<Q> {
     rawPathname: string;
     url: string;
 }
-export interface HickoryLocation<Q> extends KeylessLocation<Q> {
+export interface SessionLocation<Q> extends Location<Q> {
     key: string;
 }
-export declare type AnyLocation<Q> = HickoryLocation<Q> | PartialLocation<Q>;
+export declare type AnyLocation<Q> = SessionLocation<Q> | PartialLocation<Q>;