+++ packages/root/types/types/navigationConfirmation.d.ts
@@ -1,8 +1,8 @@
-import { HickoryLocation } from "./location";
+import { SessionLocation } from "./location";
 import { Action } from "./hickory";
 export interface NavigationInfo<Q> {
-    to: HickoryLocation<Q>;
-    from: HickoryLocation<Q>;
+    to: SessionLocation<Q>;
+    from: SessionLocation<Q>;
     action: Action;
 }
 export declare type ConfirmationFunction<Q> = (info: NavigationInfo<Q>, confirm: () => void, prevent?: () => void) => void;